Make CSV import write Auditor Tasks
CSV import used to write blobs into an `auditor_todo` collection which were then translated by the client into Task objects.  We now write task objects directly into an `auditor_task` collection.

This PR looks more complicated than it should be because of the difficulty of sharing common types and JS between the main app and cloud functions.  tsconfig and tslint weren't able to reach into ../src/store/corestore.ts without deployment issues, and so I ultimately decided to do the hacky thing by copying the common types during `yarn build` of `functions/`.  If there's a better way to do this (without creating a private module/etc/etc), please LMK.

The changes:

* Extracted common types into sharedtypes.ts.  This required tweaking a bunch of imports.
* Copy that during `yarn build` into `functions/src` and .gitignore it.
* Change CSV import to write Task objects directly.
* Remove all references to the `auditor_todo` collection, though I left firebase rules alone for now so that we don't break existing clients (until this PR gets in).
