+++ src/store/corestore.ts
@@ -1,7 +1,20 @@
-import firebase from "firebase/app";
+import * as firebase from "firebase/app";
 import "firebase/auth";
 import "firebase/firestore";
 import "firebase/functions";
+import {
+  UserRole,
+  Task,
+  TaskDecision,
+  AUDITOR_TASK_COLLECTION,
+  PAYOR_TASK_COLLECTION,
+  OPERATOR_TASK_COLLECTION,
+  REJECTED_TASK_COLLECTION,
+  PAYMENT_COMPLETE_TASK_COLLECTION,
+  PaymentRecipient,
+  ACTIVE_TASK_COLLECTION,
+  removeEmptyFieldsInPlace
+} from "../sharedtypes";
 
 const FIREBASE_CONFIG = {
   apiKey: "AIzaSyCspibVcd3GcAk01xHndZEJX8zuxwPIt-Y",
@@ -11,84 +24,6 @@ const FIREBASE_CONFIG = {
   storageBucket: "flows-app-staging.appspot.com",
   messagingSenderId: "785605389839",
   appId: "1:785605389839:web:dedec19abb81b7df8a3d7a"
-};
-const AUDITOR_TODO_COLLECTION = "auditor_todo";
-const OPERATOR_TASK_COLLECTION = "operator_task";
-const PAYOR_TASK_COLLECTION = "payor_task";
-const PAYMENT_COMPLETE_TASK_COLLECTION = "payment_complete_task";
-const ACTIVE_TASK_COLLECTION = "actively_viewed_tasks";
-const REJECTED_TASK_COLLECTION = "rejected_task";
-
-type AuditorTodo = {
-  batchID: string;
-  pharmacyID: string;
-  data: any;
-};
-
-export enum UserRole {
-  AUDITOR = "Auditor",
-  PAYOR = "Payor",
-  OPERATOR = "Operator",
-  ADMIN = "Admin"
-}
-
-export enum TaskDecision {
-  DECLINE_AUDIT = "Decline Audit",
-  APPROVE_AUDIT = "Approve Audit",
-  DECLINE_PAYMENT = "Decline Payment",
-  PAYMENT_COMPLETE = "Payment Complete",
-  TASK_REJECTED = "Task Rejected"
-}
-
-export type Site = {
-  name: string;
-  phone?: string;
-};
-
-export type ClaimEntry = {
-  patientAge?: number;
-  patientFirstName: string;
-  patientLastName: string;
-  patientSex?: string;
-  patientID?: string;
-  phone?: string;
-  item: string;
-  totalCost: number;
-  claimedCost: number;
-  photoIDUri?: string;
-  photoMedUri?: string;
-  photoMedBatchUri?: string;
-  timestamp: number;
-  reviewed?: boolean;
-};
-
-export type ClaimTask = {
-  entries: ClaimEntry[];
-  site: Site;
-};
-
-export type TaskChangeMetadata = {
-  timestamp: number;
-  by: string;
-  desc?: string;
-  notes?: string;
-};
-
-export type Task = ClaimTask & {
-  id: string;
-  flow?: TaskDecision;
-  changes: TaskChangeMetadata[];
-};
-
-export type PaymentRecipient = {
-  name?: string;
-  phoneNumber: string;
-  currencyCode: string;
-  amount: number;
-  reason?: string;
-  metadata: {
-    [key: string]: any;
-  };
 };
 
 export type ActiveTask = {
@@ -118,7 +53,7 @@ export async function declineAudit(task: Task, notes?: string) {
   await saveDeclinedTask(
     task,
     TaskDecision.DECLINE_AUDIT,
-    AUDITOR_TODO_COLLECTION,
+    AUDITOR_TASK_COLLECTION,
     notes
   );
 }
@@ -192,7 +127,7 @@ export async function saveAuditorApprovedTask(
       .set(task),
     firebase
       .firestore()
-      .collection(AUDITOR_TODO_COLLECTION)
+      .collection(AUDITOR_TASK_COLLECTION)
       .doc(task.id)
       .delete()
   ]);
@@ -291,47 +226,11 @@ export function subscribeToTasks(
 }
 
 export async function loadAuditorTasks(): Promise<Task[]> {
-  const todoSnapshot = await firebase
+  const taskSnapshot = await firebase
     .firestore()
-    .collection(AUDITOR_TODO_COLLECTION)
+    .collection(AUDITOR_TASK_COLLECTION)
     .get();
-  const todos = todoSnapshot.docs.map(doc => {
-    const todo = (doc.data() as unknown) as AuditorTodo;
-    return {
-      ...todo,
-      id: doc.id
-    };
-  });
-
-  return todos
-    .filter(t => t && t.data && t.data.length > 0)
-    .map(t => {
-      const d = t.data;
-      const patients = t.data.map((d: any) => ({
-        patientAge: d["g2:A12 Age"],
-        patientFirstName: d["g2:A10 First Name"],
-        patientLastName: d["g2:A11 Last Name"],
-        patientSex:
-          d["g2:A13 Male or Female (0 male, 1 female)"] === "0" ? "M" : "F",
-        patientID: d["g4:B02"]["1 ID number on voucher"],
-        phone: d["g2:A14 Phone Number"],
-        photoIDUri: d["g4:B03"]["1 Photo of ID card"],
-        photoMedUri: d["g5:B04 (Medication)"],
-        photoMedBatchUri: d["g5:B05 (Medication batch)"],
-        item: d["Type received"],
-        totalCost: parseFloat(d["Total med price covered by SPIDER"]),
-        claimedCost: parseFloat(d["Total reimbursement"]),
-        timestamp: new Date(d["YYYY"], d["MM"] - 1, d["DD"]).getTime()
-      }));
-      return {
-        id: t.id,
-        entries: patients,
-        site: {
-          name: d[0]["g3:B01 Pharmacy name"]
-        },
-        changes: []
-      };
-    });
+  return taskSnapshot.docs.map(doc => (doc.data() as unknown) as Task);
 }
 
 export async function loadCompletedPaymentTasks(): Promise<Task[]> {
@@ -418,16 +317,4 @@ export function subscribeActiveTasks(
       onActiveTasksChanged(actives);
     });
   return unsubscriber;
-}
-
-// https://stackoverflow.com/questions/286141/remove-blank-attributes-from-an-object-in-javascript
-function removeEmptyFieldsInPlace(obj: { [key: string]: any }) {
-  Object.keys(obj).forEach(key => {
-    if (obj[key] && typeof obj[key] === "object") {
-      removeEmptyFieldsInPlace(obj[key]);
-    } else if (obj[key] == null) {
-      // Note this captures undefined as well!
-      delete obj[key];
-    }
-  });
 }
